Fix app not reading `aniyomix.extensionLib` extension metadata 2

+++ app/src/main/java/eu/kanade/tachiyomi/extension/util/ExtensionLoader.kt
@@ -248,8 +248,8 @@ internal object ExtensionLoader {
         }
 
         // Validate lib version
-        val libVersion = appInfo.metaData.getFloat(METADATA_EXTENSION_LIB)
-            .takeUnless { it == 0.0f }
+        val libVersion = appInfo.metaData.getInt(METADATA_EXTENSION_LIB)
+            .takeUnless { it == 0 }
             ?.toString()
             ?.toDouble()
             ?: versionName.substringBeforeLast('.').toDoubleOrNull()
